+++ src/main/java/net/oneandone/lavender/modules/ScmProperties.java
@@ -40,7 +40,7 @@ public class ScmProperties {
     public static final String SCM_PREFIX = "scm.";
 
     public final String name;
-    public final boolean legacy;
+    public final boolean classic;
     public final Filter filter;
     public final String connection;
 
@@ -57,7 +57,7 @@ public class ScmProperties {
     public final String source;
 
     /// CHECKSTYLE:OFF
-    public ScmProperties(String name, boolean legacy, Filter filter, String connection, String tag, String path,
+    public ScmProperties(String name, boolean classic, Filter filter, String connection, String tag, String path,
                          boolean lavendelize, String resourcePathPrefix, String targetPathPrefix, Map<String, String> indexOpt,
                          String source) {
         /// CHECKSTYLE:ON
@@ -68,7 +68,7 @@ public class ScmProperties {
             throw new IllegalArgumentException(name);
         }
         this.name = name;
-        this.legacy = legacy;
+        this.classic = classic;
         this.filter = filter;
         this.connection = connection;
         this.tag = tag;
@@ -221,7 +221,7 @@ public class ScmProperties {
 
     public String toString() {
         return "name: " + name + "\n"
-                + "legacy: " + legacy + "\n"
+                + "classic: " + classic + "\n"
                 + "tag: " + tag + "\n"
                 + "include: " + Strings.toList(filter.getIncludes()) + "\n"
                 + "exclude: " + Strings.toList(filter.getExcludes()) + "\n"